refactor(dynamic-host): 重构PPS，将业务无关代码抽离到基类BasePluginProcessService
